app/requirements.txt
openai
python-dotenv
tenacity
httpx
fastapi==0.104.1
uvicorn==0.23.2
pydantic
